Project upload
About the project: The supermarket chain Good Seed would like to explore whether Data Science can help them adhere to alcohol laws by making sure they do not sell alcohol to people underage. Use computer vision methods to determine the age of a person from a photo. Build and evaluate a model for verifying people's age with a minimum mae score of 7.

The project was identified as a regression task with one output layer and using mae as the scoring method. Exploratory data analysis was carried out with outliers removed. ResNet50 was used as the backbone in the model training with a final score of  6.7641